+++ include/intersections.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   intersections.h                                    :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: hsarhan <hsarhan@student.42.fr>            +#+  +:+       +#+        */
+/*   By: mkhan <mkhan@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/21 12:19:12 by mkhan             #+#    #+#             */
-/*   Updated: 2023/01/07 16:01:54 by hsarhan          ###   ########.fr       */
+/*   Updated: 2023/01/10 14:45:16 by mkhan            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -34,7 +34,7 @@ struct s_ray
 typedef struct s_intersect		t_intersection;
 struct s_intersect
 {	
-	float		time;
+	double		time;
 	t_shape		*shape;
 	t_vector	point;
 	t_vector	normal;
@@ -61,11 +61,11 @@ bool			intersect(t_shape *shape, const t_ray *ray,
 
 t_intersection	*hit(t_intersections *xs);
 
-void			ray_position(t_vector *pos, const t_ray *ray, float time);
+void			ray_position(t_vector *pos, const t_ray *ray, double time);
 
 t_vector		normal_at(const t_shape *shape, const t_vector *itx_point);
 
-void			ray_from_cam(t_ray *ray, const t_camera *cam, float x, float y);
+void			ray_from_cam(t_ray *ray, const t_camera *cam, double x, double y);
 bool	intersect_cylinder(const t_ray *ray, t_shape *shape,
 	t_intersections *xs);
 bool	intersect_cone(const t_ray *ray, t_shape *shape, t_intersections *xs);
