+++ libft/ft_swap.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ft_swap.c                                          :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: hsarhan <hsarhan@student.42.fr>            +#+  +:+       +#+        */
+/*   By: mkhan <mkhan@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/12/16 16:13:40 by hsarhan           #+#    #+#             */
-/*   Updated: 2022/12/23 12:15:48 by hsarhan          ###   ########.fr       */
+/*   Updated: 2023/01/10 14:45:16 by mkhan            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,9 +21,9 @@ void	ft_swap(int *a, int *b)
 	*b = tmp;
 }
 
-void	ft_swapd(float *a, float *b)
+void	ft_swapd(double *a, double *b)
 {
-	float	tmp;
+	double	tmp;
 
 	tmp = *a;
 	*a = *b;
